polish, minor refactors
