+++ garden/src/features/benchmarks/utils/matbench.ts
@@ -1,404 +1,29 @@
 // MatBench Discovery specific metadata and utilities
 // Supports both known MatBench metrics and dynamic metric discovery from backend
 
-import { BenchmarkResult } from "../api/useGetBenchmarks";
+import { BenchmarkResult } from "../types/benchmarks.types";
+import { MetricInfo, BenchmarkInfo } from "../types/benchmarks.types";
+import { MATBENCH_METRICS, MATBENCH_BENCHMARKS, ENVIRONMENT_METRIC_KEYS, EXCLUDED_METRIC_KEYS } from "./constants";
 
-export interface MetricInfo {
-  name: string;
-  description: string;
-  betterIs: 'higher' | 'lower';
-  format: 'decimal' | 'percentage' | 'integer' | 'text' | 'duration' | 'currency';
-  unit?: string;
-  decimalPlaces?: number;
-  isPrimaryMetric?: boolean;
-}
+export { MATBENCH_METRICS, MATBENCH_BENCHMARKS, ENVIRONMENT_METRIC_KEYS, EXCLUDED_METRIC_KEYS };
 
-export interface BenchmarkInfo {
-  id: string;
-  name: string;
-  description: string;
-  detailedDescription: string;
-  purpose: string;
-  learnMoreUrl?: string;
-}
+// Re-export formatMetricValue for backward compatibility during refactor
+import { formatMetricValue } from './formatting';
+export { formatMetricValue };
 
-// MatBench Discovery metric definitions
-export const MATBENCH_METRICS: Record<string, MetricInfo> = {
-  // Lowercase versions (for mock data)
-  f1_score: {
-    name: 'F1 Score',
-    description: 'Harmonic mean of precision and recall for stability prediction accuracy',
-    betterIs: 'higher',
-    format: 'decimal',
-    decimalPlaces: 3,
-    isPrimaryMetric: true,
-  },
-  daf: {
-    name: 'Discovery Acceleration Factor',
-    description: 'How many times faster this model finds stable materials vs random search',
-    betterIs: 'higher',
-    format: 'decimal',
-    decimalPlaces: 2,
-    unit: 'x faster',
-    isPrimaryMetric: true,
-  },
-  accuracy: {
-    name: 'Accuracy',
-    description: 'Percentage of correct stability predictions',
-    betterIs: 'higher',
-    format: 'percentage',
-    decimalPlaces: 1,
-  },
-  precision: {
-    name: 'Precision',
-    description: 'Of predicted stable materials, what percentage are actually stable',
-    betterIs: 'higher',
-    format: 'decimal',
-    decimalPlaces: 3,
-  },
-  recall: {
-    name: 'Recall',
-    description: 'Of all stable materials, what percentage did the model find',
-    betterIs: 'higher',
-    format: 'decimal',
-    decimalPlaces: 3,
-  },
-  // Uppercase versions (for real backend data)
-  'F1': {
-    name: 'F1 Score',
-    description: 'Harmonic mean of precision and recall for stability prediction accuracy',
-    betterIs: 'higher',
-    format: 'decimal',
-    decimalPlaces: 3,
-    isPrimaryMetric: true,
-  },
-  'DAF': {
-    name: 'Discovery Acceleration Factor',
-    description: 'How many times faster this model finds stable materials vs random search',
-    betterIs: 'higher',
-    format: 'decimal',
-    decimalPlaces: 2,
-    unit: 'x faster',
-    isPrimaryMetric: true,
-  },
-  'Accuracy': {
-    name: 'Accuracy',
-    description: 'Percentage of correct stability predictions',
-    betterIs: 'higher',
-    format: 'decimal',
-    decimalPlaces: 3,
-  },
-  'Precision': {
-    name: 'Precision',
-    description: 'Of predicted stable materials, what percentage are actually stable',
-    betterIs: 'higher',
-    format: 'decimal',
-    decimalPlaces: 3,
-  },
-  'Recall': {
-    name: 'Recall',
-    description: 'Of all stable materials, what percentage did the model find',
-    betterIs: 'higher',
-    format: 'decimal',
-    decimalPlaces: 3,
-  },
-  'MAE': {
-    name: 'Mean Absolute Error',
-    description: 'Average absolute difference between predicted and actual values',
-    betterIs: 'lower',
-    format: 'decimal',
-    decimalPlaces: 4,
-  },
-  'RMSE': {
-    name: 'Root Mean Square Error',
-    description: 'Square root of the average squared differences between predicted and actual values',
-    betterIs: 'lower',
-    format: 'decimal',
-    decimalPlaces: 4,
-  },
-  'R2': {
-    name: 'R² Score',
-    description: 'Coefficient of determination - how well the model explains variance in the data',
-    betterIs: 'higher',
-    format: 'decimal',
-    decimalPlaces: 3,
-  },
-  'TPR': {
-    name: 'True Positive Rate',
-    description: 'Sensitivity - proportion of actual positives correctly identified',
-    betterIs: 'higher',
-    format: 'decimal',
-    decimalPlaces: 3,
-  },
-  'FPR': {
-    name: 'False Positive Rate',
-    description: 'Proportion of actual negatives incorrectly identified as positive',
-    betterIs: 'lower',
-    format: 'decimal',
-    decimalPlaces: 3,
-  },
-  'TNR': {
-    name: 'True Negative Rate',
-    description: 'Specificity - proportion of actual negatives correctly identified',
-    betterIs: 'higher',
-    format: 'decimal',
-    decimalPlaces: 3,
-  },
-  'FNR': {
-    name: 'False Negative Rate',
-    description: 'Proportion of actual positives incorrectly identified as negative',
-    betterIs: 'lower',
-    format: 'decimal',
-    decimalPlaces: 3,
-  },
-  // Confusion matrix values
-  'TP': {
-    name: 'True Positives',
-    description: 'Number of correctly predicted stable materials',
-    betterIs: 'higher',
-    format: 'integer',
-    decimalPlaces: 0,
-  },
-  'TN': {
-    name: 'True Negatives',
-    description: 'Number of correctly predicted unstable materials',
-    betterIs: 'higher',
-    format: 'integer',
-    decimalPlaces: 0,
-  },
-  'FP': {
-    name: 'False Positives',
-    description: 'Number of incorrectly predicted stable materials (Type I errors)',
-    betterIs: 'lower',
-    format: 'integer',
-    decimalPlaces: 0,
-  },
-  'FN': {
-    name: 'False Negatives',
-    description: 'Number of incorrectly predicted unstable materials (Type II errors)',
-    betterIs: 'lower',
-    format: 'integer',
-    decimalPlaces: 0,
-  },
-  'R^2': {
-    name: 'R² Score',
-    description: 'Coefficient of determination - how well the model explains variance in the data',
-    betterIs: 'higher',
-    format: 'decimal',
-    decimalPlaces: 3,
-  },
-  'RMSD': {
-    name: 'RMSD',
-    description: 'Root Mean Square Displacement for structure optimization accuracy',
-    betterIs: 'lower',
-    format: 'decimal',
-    decimalPlaces: 3,
-    unit: 'Å',
-  },
-  // S2EF and S2EFS task metrics - Energy predictions
-  'energy_r2': {
-    name: 'Energy R²',
-    description: 'How well the model predicts total energy (coefficient of determination)',
-    betterIs: 'higher',
-    format: 'decimal',
-    decimalPlaces: 4,
-    isPrimaryMetric: true,
-  },
-  'energy_mae': {
-    name: 'Energy MAE',
-    description: 'Mean absolute error of energy predictions',
-    betterIs: 'lower',
-    format: 'decimal',
-    decimalPlaces: 4,
-    unit: 'eV',
-  },
-  'energy_rmse': {
-    name: 'Energy RMSE',
-    description: 'Root mean square error of energy predictions',
-    betterIs: 'lower',
-    format: 'decimal',
-    decimalPlaces: 4,
-    unit: 'eV',
-  },
-  // S2EF and S2EFS task metrics - Force predictions
-  'force_r2': {
-    name: 'Force R²',
-    description: 'How well the model predicts atomic forces (coefficient of determination)',
-    betterIs: 'higher',
-    format: 'decimal',
-    decimalPlaces: 4,
-    isPrimaryMetric: true,
-  },
-  'force_mae': {
-    name: 'Force MAE',
-    description: 'Mean absolute error of force predictions',
-    betterIs: 'lower',
-    format: 'decimal',
-    decimalPlaces: 4,
-    unit: 'eV/Å',
-  },
-  'force_rmse': {
-    name: 'Force RMSE',
-    description: 'Root mean square error of force predictions',
-    betterIs: 'lower',
-    format: 'decimal',
-    decimalPlaces: 4,
-    unit: 'eV/Å',
-  },
-  // S2EF and S2EFS task metrics - Stress predictions
-  'stress_r2': {
-    name: 'Stress R²',
-    description: 'How well the model predicts stress tensor (coefficient of determination)',
-    betterIs: 'higher',
-    format: 'decimal',
-    decimalPlaces: 4,
-  },
-  'stress_mae': {
-    name: 'Stress MAE',
-    description: 'Mean absolute error of stress predictions',
-    betterIs: 'lower',
-    format: 'decimal',
-    decimalPlaces: 4,
-    unit: 'eV/Å³',
-  },
-  'stress_rmse': {
-    name: 'Stress RMSE',
-    description: 'Root mean square error of stress predictions',
-    betterIs: 'lower',
-    format: 'decimal',
-    decimalPlaces: 4,
-    unit: 'eV/Å³',
-  },
-  // Dataset info
-  // Dataset info
-  // (Removed as per user request)
-  // ============================================================================
-  // Run Metadata & Hardware Metrics
-  // ============================================================================
-  // Hardware Info
-  'device_type': {
-    name: 'Device Type',
-    description: 'The type of compute device used: CUDA (NVIDIA GPU), MPS (Apple Silicon), or CPU',
-    betterIs: 'higher',
-    format: 'text',
-    decimalPlaces: 0,
-  },
-  'num_gpus': {
-    name: 'GPUs',
-    description: 'Total number of GPUs used for this benchmark run. More GPUs can improve throughput.',
-    betterIs: 'higher',
-    format: 'integer',
-    decimalPlaces: 0,
-  },
-  // Timing Metrics
-  'total_seconds': {
-    name: 'Runtime',
-    description: 'Total wall-clock time to complete the benchmark. Includes model loading and all structure evaluations.',
-    betterIs: 'lower',
-    format: 'duration',
-    decimalPlaces: 1,
-    unit: 'seconds',
-  },
-  'throughput_per_second': {
-    name: 'Throughput',
-    description: 'Number of structures evaluated per second. Higher throughput means faster inference.',
-    betterIs: 'higher',
-    format: 'decimal',
-    decimalPlaces: 3,
-    unit: 'structures/sec',
-    isPrimaryMetric: true,
-  },
-  'num_workers': {
-    name: 'Number of Workers',
-    description: 'Number of parallel workers used for data loading and processing.',
-    betterIs: 'higher',
-    format: 'integer',
-    decimalPlaces: 0,
-  },
-  // Cost Metrics
-  'total_gpu_hours': {
-    name: 'GPU Hrs',
-    description: 'Total GPU compute time consumed (runtime × number of GPUs). Key metric for compute budget.',
-    betterIs: 'lower',
-    format: 'decimal',
-    decimalPlaces: 4,
-    unit: 'GPU-hours',
-  },
-  'estimated_cost_usd': {
-    name: 'Cost ($)',
-    description: 'Estimated cloud compute cost based on GPU hours and typical cloud GPU pricing.',
-    betterIs: 'lower',
-    format: 'currency',
-    decimalPlaces: 4,
-    unit: 'USD',
-  },
-  'estimated_cost_per_1000_structures_usd': {
-    name: '$/1k Structs',
-    description: 'Estimated cost to evaluate 1,000 structures. Useful for comparing efficiency across different dataset sizes.',
-    betterIs: 'lower',
-    format: 'currency',
-    decimalPlaces: 4,
-    unit: 'USD/1K',
-    isPrimaryMetric: true,
-  },
-  // (Removed as per user request)
-  // Dataset Metrics
-  // (Removed as per user request to hide these metrics)
-  // Legacy lowercase for compatibility
-  rmsd: {
-    name: 'RMSD',
-    description: 'Root Mean Square Displacement for structure optimization accuracy',
-    betterIs: 'lower',
-    format: 'decimal',
-    decimalPlaces: 3,
-    unit: 'Å',
-  },
-  latency_ms: {
-    name: 'Latency',
-    description: 'Average time to make a prediction',
-    betterIs: 'lower',
-    format: 'integer',
-    unit: 'ms',
-  },
-  thermal_conductivity_mae: {
-    name: 'Thermal Conductivity MAE',
-    description: 'Mean absolute error for thermal conductivity predictions',
-    betterIs: 'lower',
-    format: 'decimal',
-    decimalPlaces: 3,
-    unit: 'W/mK',
-  },
-};
+// Get practical impact message for a model's performance
+export const getPracticalImpact = (_f1Score?: number, daf?: number): string => {
+  const acceleration = daf || 1;
 
-// Helper to get display name for benchmark
-export const getBenchmarkDisplayName = (benchmarkName: string): string => {
-  // Check exact match in definitions
-  if (MATBENCH_BENCHMARKS[benchmarkName]) {
-    return MATBENCH_BENCHMARKS[benchmarkName].name;
+  if (acceleration >= 5) {
+    return `Could save years of lab work by screening materials ${acceleration.toFixed(1)}x faster`;
+  } else if (acceleration >= 3) {
+    return `Significantly accelerates materials discovery (${acceleration.toFixed(1)}x faster)`;
+  } else if (acceleration >= 2) {
+    return `Modest acceleration for materials screening (${acceleration.toFixed(1)}x faster)`;
+  } else {
+    return `Limited acceleration over random search (${acceleration.toFixed(1)}x)`;
   }
-
-  // Check for common variations (underscore vs hyphen)
-  const normalized = benchmarkName.replace(/_/g, '-');
-  if (MATBENCH_BENCHMARKS[normalized]) {
-    return MATBENCH_BENCHMARKS[normalized].name;
-  }
-
-  // Fallback: Title Case
-  return benchmarkName
-    .replace(/[_-]/g, ' ')
-    .replace(/\b\w/g, c => c.toUpperCase());
-};
-
-// MatBench Discovery benchmark definitions
-export const MATBENCH_BENCHMARKS: Record<string, BenchmarkInfo> = {
-  'matbench-discovery': {
-    id: 'matbench-discovery',
-    name: 'MatBench Discovery',
-    description: 'Evaluates ML models for predicting stable inorganic crystal structures',
-    detailedDescription: 'MatBench Discovery simulates high-throughput discovery of new stable inorganic crystals. Models are evaluated on their ability to predict which crystal structures will be thermodynamically stable, helping accelerate materials discovery.',
-    purpose: 'Find stable materials faster than traditional methods',
-    learnMoreUrl: 'https://matbench-discovery.materialsproject.org/',
-  },
 };
 
 // Get performance tier based on F1 score and DAF
@@ -437,51 +62,23 @@ export const getPerformanceTier = (f1Score?: number, daf?: number): {
   }
 };
 
-// Format metric value for display
-export const formatMetricValue = (value: unknown, metricKey: string): string => {
-  if (value === undefined || value === null) return 'N/A';
-
-  const metric = MATBENCH_METRICS[metricKey];
-  if (!metric) return String(value);
-
-  let numValue: number;
-
-  // Handle complex objects with parsedValue
-  if (value && typeof value === 'object' && 'parsedValue' in value) {
-    numValue = (value as { parsedValue: number }).parsedValue;
-  } else if (typeof value === 'number') {
-    numValue = value;
-  } else {
-    return String(value);
+// Helper to get display name for benchmark
+export const getBenchmarkDisplayName = (benchmarkName: string): string => {
+  // Check exact match in definitions
+  if (MATBENCH_BENCHMARKS[benchmarkName]) {
+    return MATBENCH_BENCHMARKS[benchmarkName].name;
   }
 
-  if (isNaN(numValue)) return 'N/A';
-
-  switch (metric.format) {
-    case 'percentage':
-      return `${(numValue * 100).toFixed(metric.decimalPlaces || 1)}%`;
-    case 'integer':
-      return numValue.toFixed(0);
-    case 'decimal':
-      return numValue.toFixed(metric.decimalPlaces || 3);
-    default:
-      return numValue.toFixed(3);
+  // Check for common variations (underscore vs hyphen)
+  const normalized = benchmarkName.replace(/_/g, '-');
+  if (MATBENCH_BENCHMARKS[normalized]) {
+    return MATBENCH_BENCHMARKS[normalized].name;
   }
-};
 
-// Get practical impact message for a model's performance
-export const getPracticalImpact = (_f1Score?: number, daf?: number): string => {
-  const acceleration = daf || 1;
-
-  if (acceleration >= 5) {
-    return `Could save years of lab work by screening materials ${acceleration.toFixed(1)}x faster`;
-  } else if (acceleration >= 3) {
-    return `Significantly accelerates materials discovery (${acceleration.toFixed(1)}x faster)`;
-  } else if (acceleration >= 2) {
-    return `Modest acceleration for materials screening (${acceleration.toFixed(1)}x faster)`;
-  } else {
-    return `Limited acceleration over random search (${acceleration.toFixed(1)}x)`;
-  }
+  // Fallback: Title Case
+  return benchmarkName
+    .replace(/[_-]/g, ' ')
+    .replace(/\b\w/g, c => c.toUpperCase());
 };
 
 // Check if a benchmark is MatBench Discovery related
@@ -706,11 +303,6 @@ export const transformResultsForDisplay = (
   });
 };
 
-export const ENVIRONMENT_METRIC_KEYS = [
-  'device_type', 'num_gpus', 'total_seconds', 'throughput_per_second',
-  'total_gpu_hours', 'estimated_cost_usd', 'estimated_cost_per_1000_structures_usd',
-  'num_structures_processed', 'num_structures_total', 'num_workers', 'gpu_names'
-];
 
 /**
  * Extract unique metric keys from transformed display data.
@@ -719,7 +311,7 @@ export const ENVIRONMENT_METRIC_KEYS = [
 export const getDisplayMetricKeys = (displayData: Record<string, unknown>[]): string[] => {
   const metadataKeys = [
     'id', 'benchmark_name', 'benchmark_task_name', 'timestamp', 'model_name', 'model_packages',
-    ...ENVIRONMENT_METRIC_KEYS
+    ...EXCLUDED_METRIC_KEYS
   ];
 
   const keys = new Set<string>();
